🔄 synced local 'skyvern/' with remote 'skyvern/'
<!-- ELLIPSIS_HIDDEN -->

> [!IMPORTANT]
> Refactor `ActionType` into a separate file `action_types.py` and update all references across the codebase.
>
>   - **Refactoring**:
>     - Move `ActionType` from `actions.py` to new file `action_types.py`.
>     - Update imports and references to `ActionType` in `cloud/__init__.py`, `scripts/replay_task_actions.py`, and `agent.py`.
>   - **Code Organization**:
>     - `action_types.py` now contains the `ActionType` class, previously in `actions.py`.
>   - **Misc**:
>     - Remove `ActionType` definition from `actions.py`.
>
> <sup>This description was created by </sup>[<img alt="Ellipsis" src="https://img.shields.io/badge/Ellipsis-blue?color=175173">](https://www.ellipsis.dev?ref=Skyvern-AI%2Fskyvern-cloud&utm_source=github&utm_medium=referral)<sup> for fa93e2e7777876d5885b58a90bc609ccbddd168b. You can [customize](https://app.ellipsis.dev/Skyvern-AI/settings/summaries) this summary. It will automatically update as commits are pushed.</sup>

<!-- ELLIPSIS_HIDDEN -->
